refactor(BA-5814): narrow scope to AppConfigPolicy only — Fragment slice moves to BA-5827
The combined Fragment + Policy foundation PR was too large for
review. Split into two stacked PRs:

- BA-5814 (this PR) keeps just the AppConfigPolicy slice — schema,
  ORM, data types, repository pair, adapter skeleton.
- BA-5827 stacks on top with the AppConfigFragment slice; the
  Fragment migration FKs to `app_config_policies.config_name` so
  the Policy table must exist first.

This PR removes everything Fragment-related:
- Migration shrunk to `app_config_policies` only and renamed to
  `5df264862995_add_app_config_policies.py`.
- Drop `data/app_config_fragment/`, `data/app_config/`,
  `models/app_config_fragment/`, `repositories/app_config_fragment/`,
  `api/adapters/app_config_fragment.py`, and the matching tests.
- Repositories container and Adapters registry no longer reference
  the Fragment side.
