+++ src/ai/backend/manager/api/adapters/registry.py
@@ -5,7 +5,6 @@ from __future__ import annotations
 from typing import TYPE_CHECKING
 
 from ai.backend.manager.api.adapters.agent.adapter import AgentAdapter
-from ai.backend.manager.api.adapters.app_config_fragment import AppConfigFragmentAdapter
 from ai.backend.manager.api.adapters.app_config_policy import AppConfigPolicyAdapter
 from ai.backend.manager.api.adapters.artifact.adapter import ArtifactAdapter
 from ai.backend.manager.api.adapters.artifact_registry.adapter import ArtifactRegistryAdapter
@@ -74,7 +73,6 @@ class Adapters:
     def __init__(
         self,
         agent: AgentAdapter,
-        app_config_fragment: AppConfigFragmentAdapter,
         app_config_policy: AppConfigPolicyAdapter,
         artifact: ArtifactAdapter,
         artifact_registry: ArtifactRegistryAdapter,
@@ -117,7 +115,6 @@ class Adapters:
         vfs_storage: VFSStorageAdapter,
     ) -> None:
         self.agent = agent
-        self.app_config_fragment = app_config_fragment
         self.app_config_policy = app_config_policy
         self.artifact = artifact
         self.artifact_registry = artifact_registry
@@ -179,7 +176,6 @@ class Adapters:
         """
         return cls(
             agent=AgentAdapter(processors),
-            app_config_fragment=AppConfigFragmentAdapter(processors),
             app_config_policy=AppConfigPolicyAdapter(processors),
             artifact=ArtifactAdapter(processors),
             artifact_registry=ArtifactRegistryAdapter(processors),
