--- src/ai/backend/manager/models/app_config_fragment/__init__.py
@@ -1,8 +1,0 @@
-from ai.backend.manager.data.app_config_fragment.types import AppConfigScopeType
-
-from .row import AppConfigFragmentRow
-
-__all__ = (
-    "AppConfigFragmentRow",
-    "AppConfigScopeType",
-)
