--- src/ai/backend/manager/repositories/app_config_fragment/db_source/__init__.py
@@ -1,3 +1,0 @@
-from .db_source import AppConfigFragmentDBSource
-
-__all__ = ("AppConfigFragmentDBSource",)
