--- src/ai/backend/manager/repositories/app_config_fragment/repositories.py
@@ -1,23 +1,0 @@
-from dataclasses import dataclass
-from typing import Self
-
-from ai.backend.manager.repositories.app_config_fragment.admin_repository import (
-    AppConfigFragmentAdminRepository,
-)
-from ai.backend.manager.repositories.app_config_fragment.repository import (
-    AppConfigFragmentRepository,
-)
-from ai.backend.manager.repositories.types import RepositoryArgs
-
-
-@dataclass
-class AppConfigFragmentRepositories:
-    repository: AppConfigFragmentRepository
-    admin_repository: AppConfigFragmentAdminRepository
-
-    @classmethod
-    def create(cls, args: RepositoryArgs) -> Self:
-        return cls(
-            repository=AppConfigFragmentRepository(args.db),
-            admin_repository=AppConfigFragmentAdminRepository(args.db),
-        )
