+++ src/ai/backend/manager/repositories/repositories.py
@@ -2,9 +2,6 @@ from dataclasses import dataclass
 from typing import Self
 
 from ai.backend.manager.repositories.agent.repositories import AgentRepositories
-from ai.backend.manager.repositories.app_config_fragment.repositories import (
-    AppConfigFragmentRepositories,
-)
 from ai.backend.manager.repositories.app_config_policy.repositories import (
     AppConfigPolicyRepositories,
 )
@@ -91,7 +88,6 @@ from ai.backend.manager.repositories.vfs_storage.repositories import VFSStorageR
 @dataclass
 class Repositories:
     agent: AgentRepositories
-    app_config_fragment: AppConfigFragmentRepositories
     app_config_policy: AppConfigPolicyRepositories
     auth: AuthRepositories
     container_registry: ContainerRegistryRepositories
@@ -144,7 +140,6 @@ class Repositories:
     @classmethod
     def create(cls, args: RepositoryArgs) -> Self:
         agent_repositories = AgentRepositories.create(args)
-        app_config_fragment_repositories = AppConfigFragmentRepositories.create(args)
         app_config_policy_repositories = AppConfigPolicyRepositories.create(args)
         auth_repositories = AuthRepositories.create(args)
         container_registry_repositories = ContainerRegistryRepositories.create(args)
@@ -198,7 +193,6 @@ class Repositories:
 
         return cls(
             agent=agent_repositories,
-            app_config_fragment=app_config_fragment_repositories,
             app_config_policy=app_config_policy_repositories,
             auth=auth_repositories,
             container_registry=container_registry_repositories,
