--- tests/unit/manager/repositories/app_config_fragment/BUILD
@@ -1,1 +1,0 @@
-python_tests()
